Create task_2
you'll want to isolate the highest number of failed login attempts so you can later investigate information about the month when that highest value occurred.

You'll use the function that returns the largest numeric element from a list. Then, you'll pass this function into the print() function to display the result. This will allow you to determine which month to investigate further.