Create task_3
In this task, define a function named analyze_logins() that takes in two parameters, username and current_day_logins. Every time this function is called, it should display a message about the number of login attempts the user has made that day.

Note that the code cell will contain only a function definition, so running it will not produce an output.