Create task_7
You'll continue working with the analyze_logins() function and add a return statement to it. Return statements allow you to send information back to the function call.

In this task, use the return keyword to output the login_ratio from the function, so that it can be used later in your work.

You'll call the function with the same arguments used in the previous task and store the output from the function call in a variable named login_analysis. You'll then use a print() statement to display the saved information.